+++ .gitignore
@@ -104,7 +104,6 @@ ENV/
 ### VirtualEnv template
 # Virtualenv
 # http://iamzed.com/2009/05/07/a-primer-on-virtualenv/
-.Python
 [Ii]nclude
 [Ll]ib
 [Ll]ib64
@@ -175,3 +174,6 @@ tests/20*
 !tests/terraform/graph/variable_rendering/test_resources/tfvar_module_variables/modules/instance
 tests/common/runner_registry/packages_csv_results/
 tests/console
+
+# sast go mod
+checkov/sast_core/vendor
