+++ checkov/sast_core/go.mod
@@ -1,3 +1,5 @@
 module github.com/bridgecrewio/checkov/checkov/sast_core
 
 go 1.19
+
+require github.com/smacker/go-tree-sitter v0.0.0-20230226123037-c459dbde1464
